+++ tm-tool/lapg/src/org/textmapper/lapg/lalr/ParserTables.java
@@ -26,15 +26,15 @@ class ParserTables implements ParserData {
 	private Symbol[] sym;
 	private int rules, nsyms, nterms, nstates;
 	private int[] rleft, rlen;
-	private short[] sym_goto, sym_from, sym_to, action_table;
+	private int[] sym_goto, sym_from, sym_to, action_table;
 	private int[] action_index;
 	private int[] final_states;
 
 	ParserTables(Symbol[] sym,
 				 int rules, int nsyms, int nterms, int nstates,
 				 int[] rleft, int[] rlen,
-				 short[] sym_goto, short[] sym_from, short[] sym_to,
-				 short[] action_table, int[] action_index, int[] final_states) {
+				 int[] sym_goto, int[] sym_from, int[] sym_to,
+				 int[] action_table, int[] action_index, int[] final_states) {
 		this.sym = sym;
 		this.rules = rules;
 		this.nsyms = nsyms;
@@ -81,22 +81,22 @@ class ParserTables implements ParserData {
 	}
 
 	@Override
-	public short[] getSymGoto() {
+	public int[] getSymGoto() {
 		return sym_goto;
 	}
 
 	@Override
-	public short[] getSymFrom() {
+	public int[] getSymFrom() {
 		return sym_from;
 	}
 
 	@Override
-	public short[] getSymTo() {
+	public int[] getSymTo() {
 		return sym_to;
 	}
 
 	@Override
-	public short[] getLalr() {
+	public int[] getLalr() {
 		return action_table;
 	}
